Fix lint errors, add testability to static classes (#559)
This fixes a number of lint errors that were somehow committed, and updates a few of the "static-only" classes so that they are instead singletons that can have their contents mocked/faked by tests. Static only classes have a tendency to discourage testing by making it difficult to test because they can't be overridden.