Category picker (#565)
This adds a widget that will do Category picking, or picking from any other hierarchical data structure.

It is designed to be incorporated via a query function that is given a "path" of values to the location in the hierarchy, so that the entire hierarchy doesn't need to be given to the widget. It just needs to hold the path to the location. The data types for the values and the hierarchy are both generic, so that any type may be used.

It supports creating new entries in the hierarchy, deleting selected items, and selecting new items.

Here's an interactive example of the picker (if it doesn't load, try again at the top of the hour: DartPad occasionally runs out of GitHub quota).

This is a first draft of this widget, and will probably need refinement.

I added support for macos and linux in the example, only because it is easy/fast to prototype things on those platforms. We can skip that if you don't think it is useful. Consequently, all the files in this PR under packages/smooth_ui_library/example/{linux,macos} are auto-generated by flutter create, so there is no need to review them deeply.