fix: 4206 - search page misalignment in search-history_view.dart (#4210) (#4266)
* fix: 4206 - search page misalignment in search-history_view.dart (#4210)

Impacted file:

search-history_view.dart: added padding to the list tile for proper alignment
Resolved the misalignment issue on the search page by adding appropriate padding to the list tile in the search-history_view.dart file. This adjustment ensures that the elements within the list are properly aligned, improving the visual consistency of the search page.

* fix: 4206 - search page misalignment in search-history_view.dart (#4210)

Impacted file:

search-history_view.dart: added padding to the list tile for proper alignment
Resolved the misalignment issue on the search page by adding appropriate padding to the list tile in the search-history_view.dart file. This adjustment ensures that the elements within the list are properly aligned, improving the visual consistency of the search page.

* fix: Update ripple effect in search-history_view.dart

Refactored the ripple effect in the search-history_view.dart file to enhance the visual feedback when interacting with list tiles

* fix: Formatted Code

Refactored the code to use floating values in the search-history_view.dart file

* fix: Formatted Code

Refactored the code to use floating values in the search-history_view.dart file

* fixed the ripple effect in edit icon in search_history_view.dart